# PLYLoader: Improve parsing performance.