**Description**

The loader redundantly performs property mapping per vertex and face and converts entire binary files to ascii for header parsing.

Only map element properties to output geometry properties as required instead of per vertex or face.
Avoid converting entire binary files to ascii for header parsing - ( 5s before, sub 1ms for lucy.ply after changes)  

Approximate performance changes:

lucy100K.ply 

Before: 100ms
After: 50ms

lucy.ply (original 500MB Stanford copy)

Before: 27s
After: 7s