@Mugen87 2023-01-02
https://raw.githack.com/aardgoose/three.js/plyloader-perf/examples/webgl_loader_ply.html